+++ uCNC/dio_control.h
@@ -39,13 +39,13 @@ uint8_t dio_get_controls(uint8_t controlmask);
 void dio_enable_probe();
 void dio_disable_probe();
 bool dio_get_probe();
-uint16_t dio_get_inputs();
+uint32_t dio_get_inputs();
 
 //outputs
-void dio_set_outputs(uint16_t mask);
-void dio_clear_outputs(uint16_t mask);
-void dio_toogle_outputs(uint16_t mask);
-uint16_t dio_get_outputs();
+void dio_set_outputs(uint32_t mask);
+void dio_clear_outputs(uint32_t mask);
+void dio_toogle_outputs(uint32_t mask);
+uint32_t dio_get_outputs();
 
 //analogs
 uint8_t dio_get_analog(uint8_t channel);
